+++ packages/schema/jest.config.js
@@ -3,5 +3,4 @@
 module.exports = {
   preset: 'ts-jest',
   displayName: 'schema',
-  collectCoverageFrom: ['./src'],
 };