+++ packages/admin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @paljs/admin
 
+## 5.3.0
+
+### Minor Changes
+
+- Update packages, update prisma to version 4.11, update apollo server to version 4
+
+### Patch Changes
+
+- Updated dependencies
+  - @paljs/types@5.3.0
+
 ## 5.2.0
 
 ### Minor Changes