+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @paljs/types
 
+## 5.3.0
+
+### Minor Changes
+
+- Update packages, update prisma to version 4.11, update apollo server to version 4
+
 ## 5.2.0
 
 ### Minor Changes